requirements.txt
python-telegram-bot[http2]
GPUtil
requests
torch
